- Fixes the tidal basin water purification after starting Broken Steel, for real this time. - Adds dehydration restoration values to the purified basin water for hardcore players. - Fixes General Chase's dialogue so you dont get stuck at the end of anchorage.
A large portion of the TTW community, myself included, misunderstood what was supposed to happen with the water, causing people to think there was a bug when it was actually working as intended. The change has been reverted with this plugin.  The general chase thing has supposedly been fixed on like 38 seperate occasions, but it's still missing from the latest TTW master.  This plugin changes "The simulation has condluded, soldier." to only fire if you speak to him while you're being teleported out, thus eliminating any chance of getting the wrong dialogue after killing Jingwei.